+++ src/ui/mainwindow/modeldataholder.cpp
@@ -778,6 +778,8 @@ void ModelDataHolder::OptimizeAll()
 {
     for (int i = 1; i < m_modelsWidget->count(); i++) {
         if (qobject_cast<ModelWidget*>(m_modelsWidget->widget(i))) {
+            if(!m_model_widgets[i - 1]->isChecked())
+                continue;
             ModelWidget* model = qobject_cast<ModelWidget*>(m_modelsWidget->widget(i));
             model->GlobalMinimize();
         }
@@ -803,6 +805,8 @@ void ModelDataHolder::CompareReduction()
 
     QVector<QPair<QJsonObject, QVector<int>>> models;
     for (int i = 1; i < m_modelsWidget->count(); i++) {
+        if(!m_model_widgets[i - 1]->isChecked())
+            continue;
         if (qobject_cast<ModelWidget*>(m_modelsWidget->widget(i))) {
             ModelWidget* modelwidget = qobject_cast<ModelWidget*>(m_modelsWidget->widget(i));
             QJsonObject model = modelwidget->Model()->ExportModel();
@@ -837,6 +841,8 @@ void ModelDataHolder::CompareAIC()
 {
     QVector<QWeakPointer<AbstractModel>> models;
     for (int i = 1; i < m_modelsWidget->count(); i++) {
+        if(!m_model_widgets[i - 1]->isChecked())
+            continue;
         if (qobject_cast<ModelWidget*>(m_modelsWidget->widget(i))) {
             ModelWidget* modelwidget = qobject_cast<ModelWidget*>(m_modelsWidget->widget(i));
             models << modelwidget->Model();